% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_connect.R
\name{tab_connect}
\alias{tab_connect}
\title{Connect to API}
\usage{
tab_connect(server_name = NULL, api_version = NULL, user = NULL,
  pass = NULL, ssl = T, site = "", ...)
}
\arguments{
\item{server_name}{The name or IP address of your Tableau Server.}

\item{api_version}{The API version of your Tableau REST API.}

\item{user}{Your Tableau Server user name.}

\item{pass}{Your Tableau Server password.}

\item{ssl}{Logical. Does your Tableau Server instance use SSL? The default is TRUE.}

\item{site}{A specific site for a Tableau Server running multiple sites. The default is an empty string.
If the Tablau Server is only running one site, the empty string will return your default site.}

\item{...}{Additional arguments.}
}
\description{
Connect to your Tableau Server's REST API.
}
\examples{
\dontrun{
tab_connect(server_name = "my.server.com", api_version = 2.0, user = myuser, pass = mypass, ssl = F)
}

}
